# docs/swagger.yaml
definitions:
  product.Product:
    properties:
      base_price:
        type: number
      category:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      tariffed_price:
        type: number
      updated_at:
        type: string
    required:
    - base_price
    - category
    - name
    type: object
info:
  contact: {}
paths:
  /api/management/product:
    get:
      description: FindByName product
      parameters:
      - description: name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.Product'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: FindByName product
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: Create product
      parameters:
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/product.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.Product'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create product
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: UpdateBYID product
      parameters:
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/product.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.Product'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: UpdateByID product
      tags:
      - Product
swagger: "2.0"